requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
aiosqlite>=0.17.0
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
aiofiles>=0.7.0
pypdf2
youtube-transcript-api>=0.4.4
python-dotenv>=0.19.0
emails>=0.6
stripe>=2.60.0
alembic>=1.7.1
pytest>=6.2.5
httpx>=0.18.2
google-generativeai>=0.3.0 